+++ src/main/java/dao/util/ConnectionWrapper.java
@@ -1,4 +1,4 @@
-package dao;
+package dao.util;
 
 import org.apache.logging.log4j.LogManager;
 import org.apache.logging.log4j.Logger;
