+++ src/main/java/dao/util/PreparedStatementSetter.java
@@ -1,4 +1,4 @@
-package dao;
+package dao.util;
 
 import java.sql.PreparedStatement;
 import java.sql.SQLException;
